# Speed up `create_batch_from_map`