# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #338
Closes #431

To be reviewed/merged after #320



Benchmark results db-benchmark:

#320
```
q1 took 33 ms
q2 took 369 ms
q3 took 1875 ms
q4 took 46 ms
q5 took 1756 ms
q7 took 1686 ms
q10 OOM
```

This PR (~20% faster for queries with smaller groups, NO OOM)
```
q1 took 34 ms
q2 took 323 ms
q3 took 1355 ms
q4 took 49 ms
q5 took 1252 ms
q7 took 1294 ms
q10 took 9550 ms
```

 # Rationale for this change
<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->
Previously, arrays were created per-row in a inefficient way:

* There is overhead for generating the array structure for each row by Using `ScalarValue::to_array`
* The single-row arrays are concatenated afterwards at the end, which is slow and would be unnecessary if they are created immediately instead
* Intermediate Vecs are generated, causing more memory usage / allocations / fragmentation.

# What changes are included in this PR?
<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->
Using `ScalarValue::iter_to_array` to create arrays instead, removing use of most intermediate Vecs / Arrays and concatenation.
This is not as efficient as it *could* be when data was already contained in typed/contiguous memory, but should be OK for most queries, and much better than before this PR.

My view is that at some point data in aggregations should be stored in contiguous arrays and only referenced (with offsets) to from other places.

# Are there any user-facing changes?
<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->
No
<!--
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->