@alamb 2021-05-25
> My view is that at some point data in aggregations should be stored in contiguous arrays and only referenced (with offsets) to from other places.

I think this makes a lot of sense.

The reason we can't use Arrow arrays for this is that for now they are not *mutable* -- making some version of an ArrowVec would be helpful (I think I remember @ritchie46  mentioning he made something like this for polars-rs)